Update README.md
Added style tag to exported chart for Chart.js implementation

When I first tested this library out I wasn't sure if it was charting, adding the style tag in the example might help people getting started. 😁 

(There is a similar issue with UPlot but I was unable to see the charted line after adding height and width).